pyproject.toml
[tool.poetry]
name = "python-grpc-testing-demo"
version = "0.0.0"
description = ""
authors = ["b"]
readme = "README.md"
packages = [
    { include = "nikitabarskov", from = "src" },
        { include = "nikitabarskov", from = "gen" },
]

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.62.0"


[tool.poetry.group.dev.dependencies]
grpcio-testing = "^1.62.0"
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
